+++ internal/common/common.go
@@ -5,15 +5,31 @@ type TestConfig struct {
 	BastionSupported             bool
 	TeleportSupported            bool
 	ExternalDnsSupported         bool
+	CertManagerSupported         bool
 	ControlPlaneMetricsSupported bool
+	ObservabilityBundleInstalled bool
+	SecurityBundleInstalled      bool
+}
+
+func NewTestConfigWithDefaults() *TestConfig {
+	return &TestConfig{
+		AutoScalingSupported:         true,
+		BastionSupported:             false,
+		TeleportSupported:            true,
+		ExternalDnsSupported:         true,
+		CertManagerSupported:         true,
+		ControlPlaneMetricsSupported: true,
+		ObservabilityBundleInstalled: true,
+		SecurityBundleInstalled:      true,
+	}
 }
 
 func Run(cfg *TestConfig) {
-	RunApps()
+	RunApps(cfg)
 	runBasic()
-	runCertManager()
+	runCertManager(cfg.CertManagerSupported)
 	runDNS(cfg.BastionSupported)
-	runMetrics(cfg.ControlPlaneMetricsSupported)
+	runMetrics(cfg)
 	runTeleport(cfg.TeleportSupported)
 	runHelloWorld(cfg.ExternalDnsSupported)
 	runScale(cfg.AutoScalingSupported)