+++ providers/capa/china/capa_test.go
@@ -5,9 +5,9 @@ import (
 
 	. "github.com/onsi/ginkgo/v2" //nolint:staticcheck
 
-	"github.com/giantswarm/cluster-test-suites/v3/internal/common"
-	"github.com/giantswarm/cluster-test-suites/v3/internal/state"
-	"github.com/giantswarm/cluster-test-suites/v3/internal/timeout"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/common"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/state"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/timeout"
 )
 
 var _ = Describe("Common tests", func() {
@@ -16,11 +16,5 @@ var _ = Describe("Common tests", func() {
 		state.SetTestTimeout(timeout.DeployApps, time.Minute*25)
 	})
 
-	common.Run(&common.TestConfig{
-		AutoScalingSupported:         true,
-		BastionSupported:             false,
-		TeleportSupported:            true,
-		ExternalDnsSupported:         true,
-		ControlPlaneMetricsSupported: true,
-	})
+	common.Run(common.NewTestConfigWithDefaults())
 })