+++ providers/capa/standard/capa_test.go
@@ -3,18 +3,12 @@ package standard
 import (
 	. "github.com/onsi/ginkgo/v2" //nolint:staticcheck
 
-	"github.com/giantswarm/cluster-test-suites/v3/internal/common"
-	"github.com/giantswarm/cluster-test-suites/v3/internal/ecr"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/common"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/ecr"
 )
 
 var _ = Describe("Common tests", func() {
-	common.Run(&common.TestConfig{
-		AutoScalingSupported:         true,
-		BastionSupported:             false,
-		TeleportSupported:            true,
-		ExternalDnsSupported:         true,
-		ControlPlaneMetricsSupported: true,
-	})
+	common.Run(common.NewTestConfigWithDefaults())
 
 	// ECR Credential Provider specific tests
 	ecr.Run()