+++ providers/capa/upgrade/capa_test.go
@@ -3,22 +3,16 @@ package upgrade
 import (
 	. "github.com/onsi/ginkgo/v2" //nolint:staticcheck
 
-	"github.com/giantswarm/cluster-test-suites/v3/internal/common"
-	"github.com/giantswarm/cluster-test-suites/v3/internal/ecr"
-	"github.com/giantswarm/cluster-test-suites/v3/internal/upgrade"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/common"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/ecr"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/upgrade"
 )
 
 var _ = Describe("Basic upgrade test", Ordered, func() {
 	upgrade.Run(upgrade.NewTestConfigWithDefaults())
 
 	// Finally run the common tests after upgrade is completed
-	common.Run(&common.TestConfig{
-		AutoScalingSupported:         true,
-		BastionSupported:             false,
-		TeleportSupported:            true,
-		ExternalDnsSupported:         true,
-		ControlPlaneMetricsSupported: true,
-	})
+	common.Run(common.NewTestConfigWithDefaults())
 
 	// ECR Credential Provider specific tests
 	ecr.Run()