+++ providers/capv/on-capa/capv_test.go
@@ -3,17 +3,15 @@ package standard
 import (
 	. "github.com/onsi/ginkgo/v2" //nolint:staticcheck
 
-	"github.com/giantswarm/cluster-test-suites/v3/internal/common"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/common"
 )
 
 var _ = Describe("Common tests", func() {
-	common.Run(&common.TestConfig{
-		// No autoscaling on-prem
-		AutoScalingSupported: false,
-		BastionSupported:     false,
-		TeleportSupported:    true,
-		// Disabled until https://github.com/giantswarm/roadmap/issues/1037
-		ExternalDnsSupported:         false,
-		ControlPlaneMetricsSupported: true,
-	})
+	cfg := common.NewTestConfigWithDefaults()
+	// No autoscaling on-prem
+	cfg.AutoScalingSupported = false
+	// Disabled until https://github.com/giantswarm/roadmap/issues/1037
+	cfg.ExternalDnsSupported = false
+
+	common.Run(cfg)
 })