+++ providers/capv/upgrade/capv_test.go
@@ -3,21 +3,18 @@ package upgrade
 import (
 	. "github.com/onsi/ginkgo/v2" //nolint:staticcheck
 
-	"github.com/giantswarm/cluster-test-suites/v3/internal/common"
-	"github.com/giantswarm/cluster-test-suites/v3/internal/upgrade"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/common"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/upgrade"
 )
 
 var _ = Describe("Basic upgrade test", Ordered, func() {
 	upgrade.Run(upgrade.NewTestConfigWithDefaults())
 
 	// Finally run the common tests after upgrade is completed
-	common.Run(&common.TestConfig{
-		// No autoscaling on-prem
-		AutoScalingSupported: false,
-		BastionSupported:     false,
-		TeleportSupported:    true,
-		// Disabled until https://github.com/giantswarm/roadmap/issues/1037
-		ExternalDnsSupported:         false,
-		ControlPlaneMetricsSupported: true,
-	})
+	cfg := common.NewTestConfigWithDefaults()
+	// No autoscaling on-prem
+	cfg.AutoScalingSupported = false
+	// Disabled until https://github.com/giantswarm/roadmap/issues/1037
+	cfg.ExternalDnsSupported = false
+	common.Run(cfg)
 })