+++ providers/capz/private/capz_test.go
@@ -3,17 +3,14 @@ package standard
 import (
 	. "github.com/onsi/ginkgo/v2" //nolint:staticcheck
 
-	"github.com/giantswarm/cluster-test-suites/v3/internal/common"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/common"
 )
 
 var _ = Describe("Common tests", func() {
-	common.Run(&common.TestConfig{
-		// Disabled until https://github.com/giantswarm/roadmap/issues/2693
-		AutoScalingSupported: false,
-		BastionSupported:     false,
-		TeleportSupported:    true,
-		// Disabled until wildcard ingress support is added
-		ExternalDnsSupported:         false,
-		ControlPlaneMetricsSupported: true,
-	})
+	cfg := common.NewTestConfigWithDefaults()
+	// Disabled until https://github.com/giantswarm/roadmap/issues/2693
+	cfg.AutoScalingSupported = false
+	// Disabled until wildcard ingress support is added
+	cfg.ExternalDnsSupported = false
+	common.Run(cfg)
 })