+++ providers/eks/upgrade/eks_test.go
@@ -3,19 +3,26 @@ package upgrade
 import (
 	. "github.com/onsi/ginkgo/v2" //nolint:staticcheck
 
-	"github.com/giantswarm/cluster-test-suites/v3/internal/common"
-	"github.com/giantswarm/cluster-test-suites/v3/internal/upgrade"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/common"
+	"github.com/giantswarm/cluster-test-suites/v4/internal/upgrade"
 )
 
 var _ = Describe("Basic upgrade test", Ordered, func() {
-	upgrade.Run(upgrade.NewTestConfigWithDefaults())
+	cfg := upgrade.NewTestConfigWithDefaults()
+	// EKS doesn't have any of the Giant Swarm apps deployed
+	cfg.ObservabilityBundleInstalled = false
+	cfg.SecurityBundleInstalled = false
+	upgrade.Run(cfg)
 
 	// Finally run the common tests after upgrade is completed
-	common.Run(&common.TestConfig{
-		AutoScalingSupported: true,
-		BastionSupported:     false,
-		ExternalDnsSupported: true,
-		// EKS does not have metrics for k8s control plane components.
-		ControlPlaneMetricsSupported: false,
-	})
+	ccfg := common.NewTestConfigWithDefaults()
+	// EKS does not have metrics for k8s control plane components.
+	ccfg.ControlPlaneMetricsSupported = false
+	// EKS doesn't have any of the Giant Swarm apps deployed
+	ccfg.ObservabilityBundleInstalled = false
+	ccfg.SecurityBundleInstalled = false
+	ccfg.ExternalDnsSupported = false
+	ccfg.AutoScalingSupported = false
+	ccfg.CertManagerSupported = false
+	common.Run(ccfg)
 })