+++ CHANGELOG.md
@@ -7,6 +7,36 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.0.3] - 2026-02-25
+
+### Changed
+
+- Go: Update `cluster-standup-teardown` to v4.1.2.
+
+## [4.0.2] - 2026-02-18
+
+### Fixed
+
+- Adjust `eks` provider detection logic to fix the upgrade test suites.
+
+## [4.0.1] - 2026-02-17
+
+### Changed
+
+- Go: Update dependencies.
+- Go: Update `clustertest` to v3.1.3 & `cluster-standup-teardown` to v4.1.1.
+
+## [4.0.0] - 2026-02-10
+
+### Added
+
+- Add `CertManagerSupported`, `ObservabilityBundleInstalled` and `SecurityBundleInstalled` configuration options to control bundle app testing and introduce `NewTestConfigWithDefaults()` constructor for simplified test configuration.
+
+### Changed
+
+- Refactored `common.RunApps` to accept the full `TestConfig` struct instead of individual boolean parameters for better extensibility.
+- Use the `giantswarm-garm-wc-e2e` garm for WC e2e testing on `garm`
+
 ## [3.1.0] - 2026-01-22
 
 ### Changed
@@ -1305,7 +1335,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Example common tests
 - Dockerfile for running tests in CI
 
-[Unreleased]: https://github.com/giantswarm/cluster-test-suites/compare/v3.1.0...HEAD
+[Unreleased]: https://github.com/giantswarm/cluster-test-suites/compare/v4.0.3...HEAD
+[4.0.3]: https://github.com/giantswarm/cluster-test-suites/compare/v4.0.2...v4.0.3
+[4.0.2]: https://github.com/giantswarm/cluster-test-suites/compare/v4.0.1...v4.0.2
+[4.0.1]: https://github.com/giantswarm/cluster-test-suites/compare/v4.0.0...v4.0.1
+[4.0.0]: https://github.com/giantswarm/cluster-test-suites/compare/v3.1.0...v4.0.0
 [3.1.0]: https://github.com/giantswarm/cluster-test-suites/compare/v3.0.0...v3.1.0
 [3.0.0]: https://github.com/giantswarm/cluster-test-suites/compare/v2.7.1...v3.0.0
 [2.7.1]: https://github.com/giantswarm/cluster-test-suites/compare/v2.7.0...v2.7.1